restructure tests
* split up tests into separate files
* fix coverage generation for SSR
