+++ mocha.coverage.opts
@@ -2,4 +2,4 @@
 --require reify
  --recursive
 ./**/__test__.js
-test/test.js
+test/*.js
