+++ mocha.opts
@@ -1,4 +1,4 @@
 --require reify
 --recursive
 ./**/__test__.js
-test/test.js
+test/*.js
