+++ package.json
@@ -14,7 +14,7 @@
     "coverage": "nyc report --reporter=text-lcov > coverage.lcov",
     "codecov": "codecov",
     "precodecov": "npm run coverage",
-    "lint": "eslint src",
+    "lint": "eslint src test/*.js",
     "build": "npm run build:main && npm run build:ssr",
     "build:main": "rollup -c rollup.config.main.js",
     "build:ssr": "rollup -c rollup.config.ssr.js",
@@ -66,10 +66,10 @@
   },
   "nyc": {
     "include": [
-      "compiler/**/*.js"
+      "src/**/*.js"
     ],
     "exclude": [
-      "compiler/**/__test__.js"
+      "src/**/__test__.js"
     ]
   },
   "babel": {
