+++ src/server-side-rendering/compile.js
@@ -1,4 +1,4 @@
-import { parse, validate } from 'svelte';
+import { parse, validate } from '../index.js';
 import { walk } from 'estree-walker';
 import deindent from '../utils/deindent.js';
 import isReference from '../utils/isReference.js';
